Devapi2/Devapi/drf/requirements.txt
django>=4.0.0,<4.1.0
djangorestframework
pyyaml
requests
django-cors-headers
django-rest-multiple-models
python-magic